Removing mesh_gl abstract GPU data sources
