+++ plugins/mmvtkm_gl/include/mmvtkm_gl/mmvtkmStreamlineRenderer.h
@@ -1,0 +1,36 @@
+/*
+ * mmvtkmStreamlineRenderer.h
+ *
+ * Copyright (C) 2020-2021 by Universitaet Stuttgart (VISUS).
+ * All rights reserved.
+ */
+
+#ifndef MMVTKM_STREAMLINE_RENDERER_H_INCLUDED
+#define MMVTKM_STREAMLINE_RENDERER_H_INCLUDED
+
+#include "mmstd/generic/CallGeneric.h"
+
+#include "mesh_gl/BaseMeshRenderer.h"
+
+namespace megamol {
+namespace mmvtkm_gl {
+
+inline constexpr char mmvtkmstreamlinerenderer_name[] = "mmvtkmStreamlineRenderer";
+
+inline constexpr char mmvtkmstreamlinerenderer_desc[] = "Renderer for streamline meshes generated by vtkm.";
+
+class mmvtkmStreamlineRenderer
+        : public mesh_gl::BaseMeshRenderer<mmvtkmstreamlinerenderer_name, mmvtkmstreamlinerenderer_desc> {
+public:
+    mmvtkmStreamlineRenderer() = default;
+    ~mmvtkmStreamlineRenderer() = default;
+
+protected:
+    void createMaterialCollection() override;
+    void updateRenderTaskCollection(bool force_update) override;
+};
+
+} // namespace mmvtkm_gl
+} // namespace megamolm
+
+#endif // !MMVTKM_STREAMLINE_RENDERER_H_INCLUDED
