+++ plugins/mmvtkm_gl/shaders/vtkm_gl/streamline.frag.glsl
@@ -1,3 +1,5 @@
+#version 450
+
 out vec4 frag_color;
 
 in vec4 color;
