--- plugins/mmvtkm_gl/shaders/vtkm_streamline.btf
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<btf type="MegaMolGLSLShader" version="1.0" namespace="vtkm_streamline">
-    <include file="common"/>
-
-    <shader name="vertex">
-        <snippet type="version">450</snippet>
-        <snippet name="CommonDefines"    type="file">commondefines.glsl</snippet>
-        <snippet name="Main"             type="file">vtkm/streamline_vertex.glsl</snippet>
-    </shader>
-
-    <!-- geometry shader -->
-
-    <shader name="fragment">
-        <snippet type="version">450</snippet>
-        <snippet name="CommonDefines"    type="file">commondefines.glsl</snippet>
-        <snippet name="Main"             type="file">vtkm/streamline_fragment.glsl</snippet>
-    </shader>
-
-</btf>
